Feature/find friend

+++ src/main/java/com/swyp/playground/domain/findfriend/service/FindFriendService.java
@@ -2,10 +2,13 @@ package com.swyp.playground.domain.findfriend.service;
 
 import com.swyp.playground.domain.child.repository.ChildRepository;
 import com.swyp.playground.domain.findfriend.domain.FindFriend;
+import com.swyp.playground.domain.findfriend.domain.PlayHistory;
 import com.swyp.playground.domain.findfriend.domain.RecruitmentStatus;
 import com.swyp.playground.domain.findfriend.dto.*;
 import com.swyp.playground.domain.findfriend.repository.FindFriendRepository;
+import com.swyp.playground.domain.findfriend.repository.PlayHistoryRepository;
 import com.swyp.playground.domain.parent.domain.Parent;
+import com.swyp.playground.domain.parent.repository.ParentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -16,6 +19,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import java.util.Optional;
@@ -29,12 +33,14 @@ public class FindFriendService {
 
     private final FindFriendRepository findFriendRepository;
 
-    private final ChildRepository childRepository;
+    private final PlayHistoryRepository playHistoryRepository;
+
+    private final ParentRepository parentRepository;
 
 
     //놀이터 친구 모집글 목록 반환
     public List<FindFriendListResponse> getFindFriendList(String playgroundId) {
-        List<FindFriend> findFriendList = findFriendRepository.findAllByPlaygroundId(playgroundId);
+        List<FindFriend> findFriendList = findFriendRepository.findAllByPlaygroundIdOrderByCreatedAtDesc(playgroundId);
         return findFriendList.stream()
                 .map(f -> FindFriendListResponse.builder()
                         .id(f.getFindFriendId())
@@ -52,8 +58,14 @@ public class FindFriendService {
     public FindFriendInfoResponse getFindFriendInfo(Long findFriendId) {
         FindFriend findFriend = findFriendRepository.findById(findFriendId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 친구 모집글을 찾을 수 없습니다."));
-        List<Parent> participants = findFriend.getParticipants();
 
+        PlayHistory playHistory = playHistoryRepository.findByFindFriend_FindFriendId(findFriendId);
+
+        //모집이 완료된 글은 FindFriend에서 owner와 participates가 null이 되기때문에 playHistory에서 과거의 owner와 participates를 get
+        Parent owner = findFriend.getStatus() != RecruitmentStatus.COMPLETE ? findFriend.getOwner() : playHistory.getOwner();
+        List<Parent> participants = findFriend.getStatus() != RecruitmentStatus.COMPLETE ? findFriend.getParticipants() : playHistory.getParticipants();
+
+        //이미 완료된 모집글이라면 History의 정보로 찾아서 생성
         return FindFriendInfoResponse.builder()
                 .playgroundName(findFriend.getPlaygroundName())
                 .recruitmentStatus(findFriend.getStatus().name())
@@ -61,14 +73,16 @@ public class FindFriendService {
                 .description(findFriend.getDescription())
                 .scheduleTime(formatSchedule(findFriend.getStartTime(), findFriend.getEndTime()))
                 .owner(FindFriendOwnerInfoResponse.builder()
-                        .nickname(findFriend.getOwner().getNickname())
-                        .role(findFriend.getOwner().getRole().name())
-                        .address(findFriend.getOwner().getAddress())
+                        .nickname(owner.getNickname())
+                        .role(owner.getRole().name())
+                        .address(owner.getAddress())
+                        .profileImg(owner.getProfileImg())
                         .build()
                 )
-                .participants(findFriend.getParticipants().stream()
+                .participants(participants.stream()
                         .map(participant -> FindFriendParticipantsListResponse.builder()
                                 .nickname(participant.getNickname())
+                                .profileImg(participant.getProfileImg())
                                 .build()
                         )
                         .toList()
@@ -90,9 +104,15 @@ public class FindFriendService {
 
 
     //친구 모집글 등록
-    public Long registerFindFriend(String playgroundId, FindFriendRegisterRequest findFriendRegisterRequest) {
+    public Long registerFindFriend(String playgroundId, String email, FindFriendRegisterRequest findFriendRegisterRequest) {
         LocalDateTime startTime = findFriendRegisterRequest.getStartTime();
         LocalDateTime endTime = startTime.plusMinutes(findFriendRegisterRequest.getDuration());
+        Parent parent = parentRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("사용자 정보를 찾을 수 없습니다."));
+
+        Optional<FindFriend> ownerParent = findFriendRepository.findByOwner_ParentId(parent.getParentId());
+        if(ownerParent.isPresent())
+            throw new IllegalArgumentException("이미 만든 친구 모집글이 있습니다.");
 
         // FindFriend 엔티티 객체 생성
         FindFriend findFriend = FindFriend.builder()
@@ -101,19 +121,30 @@ public class FindFriendService {
                 .title(findFriendRegisterRequest.getTitle())
                 .startTime(startTime)
                 .endTime(endTime)
+                .createdAt(LocalDateTime.now())
                 .currentCount(1)
                 .status(RecruitmentStatus.RECRUITING)
                 .description(findFriendRegisterRequest.getDescription())
+                .owner(parent)
                 .build();
 
         return findFriendRepository.save(findFriend).getFindFriendId();
     }
 
     //친구 모집글 수정
-    public FindFriendInfoResponse modifyFindFriendInfo(String playgroundId, Long findFriendId, FindFriendModifyRequest findFriendModifyRequest) {
-        // 모집글 조회
+    public FindFriendInfoResponse modifyFindFriendInfo(String playgroundId, Long findFriendId, String email, FindFriendModifyRequest findFriendModifyRequest) {
+
+        Parent parent = parentRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("사용자 정보를 찾을 수 없습니다."));
+
         FindFriend findFriend = findFriendRepository.findById(findFriendId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 모집글을 찾을 수 없습니다."));
+
+
+        //친구 모집글 수정 권한 검증
+        if(findFriend.getOwner().getParentId() != parent.getParentId()){
+            throw new IllegalArgumentException("해당 글을 수정할 권한이 없습니다.");
+        }
 
         // 놀이터 ID 검증
         if (!findFriend.getPlaygroundId().equals(playgroundId)) {
@@ -140,26 +171,110 @@ public class FindFriendService {
     }
 
     //친구 모집글 삭제
-    public void deleteFindFriend(Long findFriendId) {
+    public void deleteFindFriend(Long findFriendId, String email) {
+
+        Parent parent = parentRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("사용자 정보를 찾을 수 없습니다."));
+
+        FindFriend findFriend = findFriendRepository.findById(findFriendId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 모집글을 찾을 수 없습니다."));
+
+        //친구 모집글 삭제 권한 검증
+        if(findFriend.getOwner().getParentId() != parent.getParentId()){
+            throw new IllegalArgumentException("해당 글을 삭제할 권한이 없습니다.");
+        }
         findFriendRepository.deleteById(findFriendId);
     }
 
-    //일정 시간마다 친구 모집글 확인(시작 시간에서 10분이 지나면 친구 모집글 상태를 노는중으로 변경
     // 10초마다 상태 업데이트
     @Scheduled(fixedRate = 10000)
     public void updateFindFriendStatus() {
         LocalDateTime now = LocalDateTime.now(ZoneId.of("Asia/Seoul"));
 
-        // 상태가 RECRUITING이고 startTime으로부터 10분이 지난 항목을 COMPLETE로 변경
+        // 상태가 RECRUITING이고 startTime으로부터 10분이 지난 항목을 PLAYING으로 변경
         List<FindFriend> toComplete = findFriendRepository.findByStatusAndStartTimeBefore(RecruitmentStatus.RECRUITING, now.minusMinutes(10));
         for (FindFriend findFriend : toComplete) {
             findFriend.setStatus(RecruitmentStatus.PLAYING);
         }
 
-        // 상태가 PLAYING이고 endTime이 지난 항목을 CLOSED로 변경
+        // 상태가 PLAYING이고 endTime이 지난 항목을 CLOSED로 변경 및 참여 해제
         List<FindFriend> toClose = findFriendRepository.findByStatusAndEndTimeBefore(RecruitmentStatus.PLAYING, now);
         for (FindFriend findFriend : toClose) {
+            //owner와 participants를 null로 바꾸기전 과거 내용 조회를 위해 playHistory에 저장(컬렉션 공유 때문에 새 객체로 생성)
+            List<Parent> originalList = findFriend.getParticipants();
+            List<Parent> copiedList = new ArrayList<>();
+            copiedList.addAll(originalList);
+            PlayHistory playHistory = PlayHistory.builder()
+                    .owner(findFriend.getOwner())
+                    .participants(copiedList)
+                    .findFriend(findFriend)
+                    .build();
+
+            playHistoryRepository.save(playHistory);
+
             findFriend.setStatus(RecruitmentStatus.COMPLETE);
+            List<Parent> participants = findFriend.getParticipants();
+            for (Parent parent : participants) {
+                parent.setFindFriend(null);
+            }
+            findFriend.setOwner(null);
+            findFriend.getParticipants().clear();
         }
+    }
+
+    //친구 모집글 참가 및 취소
+    public void actionFindFriend(String playgroundId, Long findFriendId, String email, String action) {
+
+        Parent parent = parentRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("사용자 정보를 찾을 수 없습니다."));
+
+        FindFriend findFriend = findFriendRepository.findById(findFriendId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 모집글을 찾을 수 없습니다."));
+
+        if(action.equals("participate")){
+            Optional<FindFriend> ownerParent = findFriendRepository.findByOwner_ParentId(parent.getParentId());
+            if(ownerParent.isPresent())
+                throw new IllegalArgumentException("이미 만든 친구 모집글이 있습니다.");
+
+            if(parent.getFindFriend() != null){
+                throw new IllegalArgumentException("이미 참가한 방이 있습니다.");
+            }
+            parent.setFindFriend(findFriend);
+            findFriend.setCurrentCount(findFriend.getCurrentCount() + 1);
+        }
+
+        if(action.equals("cancel")){
+            parent.setFindFriend(null);
+            findFriend.setCurrentCount(findFriend.getCurrentCount() - 1);
+        }
+
+    }
+
+    //내가 모집한 글들 조회
+    public List<FindFriendListResponse> getMyFindFriendList(String email) {
+        List<FindFriend> findFriendList = new ArrayList<>();
+
+        //모집중 이거나 노는중인 글 1개 조회
+        findFriendRepository.findByOwner_Email(email)
+                .ifPresent(findFriendList::add);
+
+        //완료된 글 목록 조회
+        List<PlayHistory> playHistoryList = playHistoryRepository.findByOwner_Email(email);
+        // PlayHistory에서 FindFriend를 추출하여 findFriendList에 추가
+        playHistoryList.forEach(playHistory ->
+                findFriendList.add(playHistory.getFindFriend())
+        );
+
+        return findFriendList.stream()
+                .map(f -> FindFriendListResponse.builder()
+                        .id(f.getFindFriendId())
+                        .playgroundName(f.getPlaygroundName())
+                        .title(f.getTitle())
+                        .description(f.getDescription())
+                        .scheduleTime(formatSchedule(f.getStartTime(), f.getEndTime())) //포맷팅 시간으로 변경하여 추가
+                        .recruitmentStatus(f.getStatus().name())
+                        .currentCount(f.getCurrentCount())
+                        .build())
+                .collect(Collectors.toList());
     }
 }
